studio/requirements.txt
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.2.0
asgiref==3.7.2
audioread==3.0.1
Babel==2.14.0
bitarray==2.9.0
build==1.0.3
CacheControl==0.13.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cleo==2.1.0
cmake==3.28.1
colorama==0.4.6
coloredlogs==15.0.1
crashtest==0.4.1
cryptography==41.0.7
Cython==3.0.6
decorator==5.1.1
distlib==0.3.8
dulwich==0.21.7
exceptiongroup==1.2.0
fairseq==0.12.2
faiss-cpu==1.7.3
fastjsonschema==2.19.0
ffmpy==0.3.1
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.12.2
humanfriendly==10.0
hydra-core==1.0.7
i18n==0.2
idna==3.6
importlib-metadata==7.0.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
joblib==1.3.2
keyring==24.3.0
lazy_loader==0.3
librosa==0.10.1
lit==17.0.6
llvmlite==0.41.1
lxml==4.9.3
MarkupSafe==2.1.3
monotonic-align==1.0.0
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
networkx==3.2.1
numba==0.58.1
numpy==1.23.5
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
omegaconf==2.0.6
onnxruntime==1.15.0
onnxruntime-gpu==1.15.0
packaging==23.2
pexpect==4.9.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
pooch==1.8.0
portalocker==2.8.2
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pyproject_hooks==1.0.0
pyworld==0.3.2
PyYAML==6.0.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
resampy==0.4.2
sacrebleu==2.4.0
scikit-learn==1.3.2
scipy==1.11.4
SecretStorage==3.3.3
shellingham==1.5.4
sniffio==1.3.0
soundfile==0.12.1
soxr==0.3.7
sqlparse==0.4.4
starlette==0.35.1
sympy==1.12
tabulate==0.9.0
threadpoolctl==3.2.0
tomli==2.0.1
tomlkit==0.12.3
torch==2.1.2
torchaudio==2.1.2
torchcrepe==0.0.22
tqdm==4.66.1
triton==2.1.0
trove-classifiers==2023.11.29
typing_extensions==4.9.0
urllib3==2.1.0
virtualenv==20.25.0
virtualenv-clone==0.5.7
zipp==3.17.0

encodec
funcy
transformers

aiofiles~=23.2.1
arrow~=1.2.3
aio-pika
loguru
python-dotenv~=1.0.0
tortoise-orm[asyncpg]==0.19.3
pydantic==1.10.12